+++ test/cases/option-router-true/webpack.config.js
@@ -6,9 +6,6 @@ module.exports = {
 
   output: {
     path: path.join(__dirname, 'dist/'),
-    //publicPath: 'auto', // tested: OK
-    //publicPath: () => '/', // tested: OK
-    //publicPath: () => 'http://localhost:8080/', // tested: OK
   },
 
   resolve: {
@@ -24,31 +21,24 @@ module.exports = {
     new HtmlBundlerPlugin({
       entry: {
         index: './src/views/pages/home/index.html',
-        page1: './src/views/pages/page1.html',
-        'sub/page2': './src/views/pages/sub/page2.html',
-        'sub/page3': './src/views/pages/sub/page3.html',
         'login/signup/index': './src/views/pages/login/signup/index.eta',
       },
       css: {
         filename: 'assets/css/[name].[contenthash:8].css',
       },
-      sources: [
-        {
-          tag: 'a',
-          attributes: ['href', 'data-link'],
-        },
-        {
-          tag: 'button',
-          attributes: ['href', 'data-hyperlink'],
-        },
-      ],
-      // resolve routes in custom attributes w/o router definition
+
+      // test: if router option is specified and not disabled, then add default sources options:
+      // sources: [
+      //   {
+      //     tag: 'a',
+      //     //attributes: ['href'],
+      //     //attributes: ['data-link'],
+      //   },
+      // ],
+
+      router: true,
       // router: {
-      //   //enabled: true, // OK
-      //   //enabled: false, // OK
-      //   //test: /\.html$/, // OK
-      //   //test: [/\.html$/, /\.eta$/], // OK
-      //   //rewriteIndex: false, // OK
+      //   enabled: true, // tested: OK
       // },
     }),
   ],